pyproject.toml
[project]
name = "vecinita"
version = "0.1.0"
description = "Vecinita RAG Q&A Assistant with LangGraph"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Vecinita Team"}
]
keywords = ["rag", "langgraph", "langchain", "supabase", "fastapi", "ai", "assistant"]

dependencies = [
    # Web Framework
    "fastapi",
    "uvicorn",
    # LangChain & LangGraph
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-openai",
    "langchain-groq",
    "langchain-text-splitters",
    "langchain-huggingface",
    "langgraph",
    # Database & Vector Store
    "supabase",
    "psycopg2-binary",
    # ML & Embeddings
    "sentence-transformers",
    "fastembed",
    # Document Processing
    "beautifulsoup4",
    "pypdf",
    "unstructured",
    "playwright",
    # Utilities
    "python-dotenv",
    "pydantic",
    "requests",
    "httpx",
    "langdetect",
    "tqdm",
    "gotrue>=2.0.0",
    "langchain-ollama>=1.0.0",
    # LangSmith for tracing and monitoring
    "langsmith>=0.4.56",
    # Pin onnxruntime and tokenizers to stable versions to avoid Windows compatibility issues
    # onnxruntime dev releases don't have Windows wheels
    # tokenizers pre-releases require Rust compilation
    "onnxruntime>=1.18.0,<1.24.0.dev; sys_platform == 'win32'",
    "tokenizers>=0.15.0,<0.22.0rc",
]

[project.optional-dependencies]
dev = [
    # Development Tools
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
    # Jupyter for experimentation
    "jupyter",
    "ipython",
]

visualization = [
    # For graph visualization
    "graphviz",
    "pygraphviz",
]

all = [
    "vecinita[dev,visualization]",
]

[project.scripts]
vecinita-main = "src.main:main"
vecinita-agent = "src.rag_agent:main"
vecinita-setup = "src.setup_check:main"
vecinita-example = "scripts.example_rag_usage:main"
vecinita-fastapi = "scripts.fastapi_integration_example:main"

[project.urls]
Homepage = "https://github.com/acadiagit/vecinita"
Repository = "https://github.com/acadiagit/vecinita"
Issues = "https://github.com/acadiagit/vecinita/issues"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
]
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore::DeprecationWarning:supabase._sync.client",
    "ignore::DeprecationWarning:importlib._bootstrap",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "supabase.*",
    "langchain.*",
    "langchain_community.*",
    "langchain_core.*",
    "langchain_openai.*",
    "langchain_groq.*",
    "langchain_huggingface.*",
    "langgraph.*",
    "sentence_transformers.*",
    "fastembed.*",
    "unstructured.*",
    "playwright.*",
]
ignore_missing_imports = true

[tool.uv]
# Ensure uv resolves stable versions to avoid Windows compatibility issues
# - onnxruntime dev releases don't have Windows wheels
# - tokenizers pre-releases require Rust compilation
# See: uv.toml for detailed configuration

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]
